Fixing Capybara warnings
