--- spec/helpers/bulk_edit_helper_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-describe BulkEditHelper do
-  include_examples 'page header', :bulk_edit_page_header, 'File Manager'
-end
