+++ spec/views/curation_concerns/scanned_resources/_show_actions.html.erb_spec.rb
@@ -15,9 +15,9 @@ RSpec.describe "curation_concerns/scanned_resources/_show_actions.html.erb" do
     render partial: "curation_concerns/scanned_resources/show_actions", locals: { collector: collector, editor: editor }
   end
   it "renders a reorder link" do
-    expect(rendered).to have_link I18n.t('file_manager.link_text'), file_manager_curation_concerns_scanned_resource_path(id: resource.id)
+    expect(rendered).to have_link I18n.t('file_manager.link_text'), href: file_manager_curation_concerns_scanned_resource_path(id: resource.id)
   end
   it "renders an Edit Structure link" do
-    expect(rendered).to have_link "Edit Structure", structure_curation_concerns_scanned_resource_path(id: resource.id)
+    expect(rendered).to have_link "Edit Structure", href: structure_curation_concerns_scanned_resource_path(id: resource.id)
   end
 end
